move propagation hook storage to global vars
