+++ lib/propagation/index.js
@@ -2,7 +2,8 @@
 const honeycomb = require("./honeycomb"),
   aws = require("./aws"),
   w3c = require("./w3c"),
-  util = require("./util");
+  util = require("./util"),
+  hooks = require("./hooks");
 
 exports.REQUEST_ID_HTTP_HEADER = "X-Request-ID";
 
@@ -35,7 +36,7 @@ exports.w3c = w3c;
 exports.util = util;
 
 // exported to run on require and module load with other configurations
-exports.configure = util.configure;
+exports.configure = hooks.configure;
 
-exports.parseFromRequest = util.parseFromRequest;
-exports.headersFromContext = util.headersFromContext;
+exports.parseFromRequest = hooks.parseFromRequest;
+exports.headersFromContext = hooks.headersFromContext;
